evaluaciones/evaluacion4/gestionAPP/requirements.txt
django-ninja==1.0.1
pyjwt==2.8.0